Finalize production build process
The build script outlines it pretty well, but we essentially build main.ss as a R6RS
program, which allows the compiler more opportunities to optimize as things aren't assumed
to be dynamic/prone to change at any time, as they are if you load the file through more
"normal" means like `chez main.ss` or a REPL

I will probably continue putting all the game code solely in main.ss for convenience at
the REPL (outside of a R6RS program), though theoretically I could start using (include)
to split the code apart while still bringing it together for prod builds.
Of course, anything suitable abstracted away from the game, and rarely-changing can be
placed in its own library, as is the case with coro, geom, and raylib today.

The missing final step for a true prod distribution is copying an install of Chez Scheme
into the distribution folder and arranging for it to automatically find the boot file,
but we'll leave that to when we actually have a game to release.