Rework player movement
This does away with the velocity variables and calculations based on delta time
We don't want to rely on the interframe delta time; in an STG everything actually needs to
be frame-synced and frame-dependent so that you can have reliable replays.
Bottomdragging is done by simply clamping the new values into bounds. This means that if
you're at the bottom of the screen and have down-left pressed, your down component of
movement will be dropped and lead to a decreased movement speed on the left component
versus if you had only pressed left.