Fix desync when quick-restart is used
The game loop is as follows:
1. handle input
2. game tick
3. render
4. increment frame counter

quick restart is triggered several frames after the button is pressed due to the menu
close animation, so it actually happens in step 3 instead of step 1, which causes a game
tick to be skipped when pressing quick restart. This affects both live plays and replays.

Fix this by introducing an explicit gui tick step before game tick and moving the
animation logic there.

Pausing and unpausing still desyncs things somehow, which needs further investigation.